+++ app/src/main/java/com/ts/jpc_test/Calculate.kt
@@ -1,6 +1,6 @@
 package com.ts.jpc_test
 
-class calculate
+class Calculate
 {
 
 }
